requirements.txt
requests
PyMuPDF
google-generativeai
markdown
